add tab system to news on smaller devices
